Added Navmesh and changed Enemy movement code completely
Added an extra room to the dev room scene to test navmesh system, added a script to the door that "uncarves" the navmesh when it is opened, got rid of all original movement code for enemy and it is entirely based on the navmesh system now.  When ranged enemies are within range but the player is not in the line of sight they no longer get stuck shooting at a wall but instead try to move towards the player until they are in line of sight before stopping and firing. Enemies can change speeds based on moveSpeed variable if player is or isn't in line of sight.