Changed door prefab size and code functionality
In the interactable script, instead of using an animator to move the door I changed it to MoveTowards() function. I also increased the door's size by quite a bit to fix some issues with the enemy pathfinding as well as to not allow insanely advantageous choke points for the player.