Created basis for finalized spawning system
Zones are designated with a trigger collider and a unique tag for that zone. ZoneFinder script detects which zone the player is in and changes the PlayerZone variable on the PlayerController script. This tells the SpawnManager where to spawn the enemies using a switch case and random number generator. A very basic player "barrier" was implemented to allow enemies to come through after spawning but the player can't go through.