Fixed a variety of issues
Shrunk AI Agents to help with navmesh navigation, still is acting weird when enemies try to go through a door near the player. Fixed the footsteps and got them to work fully. Turned off the shadows the glass areas casted because of some weird dot artifacts appearing. Fixed ranged enemy and their fire point, moved enemythirdcontroller script to the scripts folder. Removed the audio source from the player prefab as well as the audio variables in the playercontroller script, replaced these with the audio controller instance. commented out the Stop() functions in the footsteps indoor and outdoor parts of the AudioController script.