Adjusted difficulty and fixed player being able to "climb" small enemies
Made a SnapToFloor function in the playercontroller script that makes the players y position 0 when he goes at or above .8, this makes it so the player can no longer climb the small enemies awkwardly. Also adjusted multiple speed and health variables in both enemy controller scripts to adjust difficulty, and reduced the amount of enemies there are at first when playing.