Added a Flashlight and Game Start functionality
Added a Flashlight script, a flashlight tag for use with the Interactable script, and different functions that don't let the game start until the player has picked up the flashlight in the map. Also made it where player's flashlight turns on automatically the first time when the flashlight in the map is picked up. This also has the dumb meta files.