Various fixes
Deleted the SwitchWeapons and weaponScript scripts in the main Assets folder and kept the ones in the scripts folder. This also broke the prefab for the UI canvas so I fixed that. Also added an extra line in the SwitchWeapons script so that the shotgun is inactive at the start of the game and there isnt 2 guns at the same time. Saved the UI canvas prefab with the new changes/fixes.