Several fixes made to broken systems
Fixed interactable script so the ammo station gives player ammo for the gun they are currently holding rather than always and only the first weapon. Made multiple changes in player controller to fix both the perk damage boost and the pick up damage boost, now both weapons will maintain the correct damage during power ups and with the perk and will also all correctly change color to red and back to normal once the powerup is over. Changed an issue where the damage dealt powerup kept one weapon's damage and applied it to all weapons while powerup was active. Fixed an issue where the gun the player was holding would not disappear from the UI on the game over screen. Attempted to fix the UI for the gun not showing up at the start of the match, but could not get it to work. Fixed the TakeDamage(damageDealt) areas of the player controller script, lines 190 and 195, so they correctly do the damage the player should do when they have power ups and/or damage perk.